requirements.txt
requests
notion-client
github-heatmap
python-dotenv
openai
python-frontmatter
Markdown